try fix duplicated code
