+++ volmdlr/utils/common_operations.py
@@ -9,6 +9,7 @@ import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 from scipy.optimize import least_squares
+import scipy.integrate as scipy_integrate
 
 import volmdlr.core
 from volmdlr.core import EdgeStyle
@@ -226,3 +227,32 @@ def get_point_distance_to_edge(edge, point, start, end):
         if math.isclose(abscissa1, abscissa2, abs_tol=1e-6):
             break
     return distance
+
+
+def ellipse_abscissa_angle_integration(ellipse3d, point_abcissa, angle_start, initial_angle):
+    """
+    Caculates the angle for a given abcissa point by integrating the ellipse.
+
+    :param ellipse3d: the Ellipse3D.
+    :param point_abcissa: the given abscissa for given point.
+    :param angle_start: Ellipse3D / ArcEllipse3D start angle. (0 for Ellipse3D).
+    :param initial_angle: angle abscissa's initial value.
+    :return: final angle abscissa's value.
+    """
+    def ellipse_arc_length(theta):
+        return math.sqrt((ellipse3d.major_axis ** 2) * math.sin(theta) ** 2 +
+                         (ellipse3d.minor_axis ** 2) * math.cos(theta) ** 2)
+
+    iter_counter = 0
+    while True:
+        res, _ = scipy_integrate.quad(ellipse_arc_length, angle_start, initial_angle)
+        if math.isclose(res, point_abcissa, abs_tol=1e-8):
+            abscissa_angle = initial_angle
+            break
+        if res > point_abcissa:
+            increment_factor = (abs(initial_angle - angle_start) * (point_abcissa - res)) / (2 * abs(res))
+        else:
+            increment_factor = (abs(initial_angle - angle_start) * (point_abcissa - res)) / abs(res)
+        initial_angle += increment_factor
+        iter_counter += 1
+    return abscissa_angle
